Test splitting repo apart, starting with react-vis-sankey
