+++ package.json
@@ -41,7 +41,6 @@
     "d3-geo": "^1.6.4",
     "d3-hierarchy": "^1.1.4",
     "d3-interpolate": "^1.1.4",
-    "d3-sankey": "^0.7.1",
     "d3-scale": "^1.0.5",
     "d3-shape": "^1.1.0",
     "d3-voronoi": "^1.1.2",
