+++ src/plot/series/canvas-wrapper.js
@@ -22,8 +22,7 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 
 import {interpolate} from 'd3-interpolate';
-import {extractAnimatedPropValues} from 'animation';
-import {ANIMATED_SERIES_PROPS} from 'utils/series-utils';
+import {ANIMATED_SERIES_PROPS, extractAnimatedPropValues} from 'animation';
 
 const MAX_DRAWS = 30;
