+++ src/plot/series/heatmap-series.js
@@ -21,7 +21,6 @@
 import React from 'react';
 
 import Animation from 'animation';
-import {ANIMATED_SERIES_PROPS} from 'utils/series-utils';
 
 import AbstractSeries from './abstract-series';
 
@@ -43,7 +42,7 @@ class HeatmapSeries extends AbstractSeries {
     }
     if (animation) {
       return (
-        <Animation {...this.props} animatedProps={ANIMATED_SERIES_PROPS}>
+        <Animation {...this.props} animatedProps={Animation.ANIMATED_SERIES_PROPS}>
           <HeatmapSeries {...this.props} animation={null}/>
         </Animation>
       );
