+++ src/plot/series/label-series.js
@@ -22,7 +22,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import AbstractSeries from './abstract-series';
 import Animation from 'animation';
-import {ANIMATED_SERIES_PROPS} from 'utils/series-utils';
 const predefinedClassName = 'rv-xy-plot__series rv-xy-plot__series--label';
 
 class LabelSeries extends AbstractSeries {
@@ -48,7 +47,7 @@ class LabelSeries extends AbstractSeries {
 
     if (animation) {
       return (
-        <Animation {...this.props} animatedProps={ANIMATED_SERIES_PROPS}>
+        <Animation {...this.props} animatedProps={Animation.ANIMATED_SERIES_PROPS}>
           <LabelSeries {...this.props} animation={null} _data={data}/>
         </Animation>
       );
