+++ src/theme.js
@@ -66,3 +66,9 @@ export const DEFAULT_SIZE = 5;
 export const DEFAULT_COLOR = DISCRETE_COLOR_RANGE[0];
 
 export const DEFAULT_TICK_SIZE = 7;
+
+export default {
+  DISCRETE_COLOR_RANGE,
+  EXTENDED_DISCRETE_COLOR_RANGE,
+  CONTINUOUS_COLOR_RANGE
+};
