+++ tests/index.js
@@ -59,7 +59,6 @@ import './components/radial-tests';
 import './components/radar-chart-tests';
 import './components/rect-series-tests';
 import './components/treemap-tests';
-import './components/sankey-tests';
 import './components/showcase-example-tests';
 import './components/sunburst-tests';
 import './components/voronoi-tests';
